change serializer

+++ main/views.py
@@ -1,6 +1,8 @@
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
+from rest_framework.decorators import APIView
+from django.http.response import Http404
 
 from Authentication.views import UserAuthentication
 from Authentication.serializer import UserSerializer
@@ -11,7 +13,6 @@ from .serializers import SubjectSerializer
 
 from project.utilis import defualtResponse, getDataFromPaginator
 
-from rest_framework.decorators import APIView
 
 # Create your views here.
 
@@ -31,6 +32,7 @@ class SubjectView(APIView):
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+    # get all subjects
     def get(self, request):
         token_ = UserAuthentication.get_token_or_none(request)
         if not token_:
@@ -40,6 +42,18 @@ class SubjectView(APIView):
         allData = getDataFromPaginator(request, subjects)
         return defualtResponse(allData, SubjectSerializer)
 
+    # get specific subject
+    @staticmethod
+    @api_view(["GET"])
+    def getSpecificSubject(request, ref):
+        token_ = UserAuthentication.get_token_or_none(request)
+        if not token_:
+            return Response({"error": "you must authorized"}, status=status.HTTP_401_UNAUTHORIZED)
+
+        subject = self.get_object(ref)
+        serializer = SubjectSerializer(subject)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
     def get_object(self, ref):
         try:
             return Subject.objects.get(ref=ref)
@@ -47,10 +61,14 @@ class SubjectView(APIView):
             raise Http404
 
     def patch(self, request, ref):
+        token_ = UserAuthentication.get_token_or_none(request)
+        if not token_:
+            return Response({"error": "you must authorized"}, status=status.HTTP_401_UNAUTHORIZED)
         subject = self.get_object(ref)
         data = request.data.copy()
         data["student"] = token_.user_id
-        serializer = SubjectSerializer(data, instance=subject)
+        print(subject)
+        serializer = SubjectSerializer(data=data, instance=subject, partial=True)
         if serializer.is_valid():
             subject = serializer.save()
             serializer = SubjectSerializer(subject)
@@ -59,49 +77,15 @@ class SubjectView(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, ref):
+        token_ = UserAuthentication.get_token_or_none(request)
+        if not token_:
+            return Response({"error": "you must authorized"}, status=status.HTTP_401_UNAUTHORIZED)
         subject = self.get_object(ref)
         try:
             subject.delete()
         except:
             return Response({"error":"error while deleting subject"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
         return Response(status=status.HTTP_200_OK)
-
-
-# @api_view(["GET", "POST"])
-# def registerSubjects(request):
-#     token_ = UserAuthentication.get_token_or_none(request)
-#     if not token_:
-#         return Response({"error": "you must authorized"}, status=status.HTTP_401_UNAUTHORIZED)
-#
-#     if request.method == "GET":
-#         subjects = Subject.objects.filter(student_id=token_.user_id).all()
-#         allData = getDataFromPaginator(request, subjects)
-#         return defualtResponse(allData, SubjectSerializer)
-#
-#     elif request.method == "POST":
-#         data = request.data.copy()
-#         data["student"] = token_.user_id
-#         serializer = SubjectSerializer(data=data)
-#         if serializer.is_valid():
-#             subject = serializer.save()
-#             serializer = SubjectSerializer(subject)
-#             return Response(serializer.data, status=status.HTTP_200_OK)
-#         else:
-#             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-#
-#     elif request.method == "PATCH":
-#         data = request.data.copy()
-#         data["student"] = token_.user_id
-#         serializer = SubjectSerializer(data=data)
-#         if serializer.is_valid():
-#             subject = serializer.save()
-#             serializer = SubjectSerializer(subject)
-#             return Response(serializer.data, status=status.HTTP_200_OK)
-#         else:
-#             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-
 
 
 @api_view(["GET"])
